wip: prperty unit store

+++ routes/v1/property_unit.php
@@ -9,9 +9,9 @@ Route::prefix('property-unit')
     ->middleware('auth:api')
     ->controller(PropertyUnitController::class)
     ->group(function () {
-        Route::get('/', 'index')->name('index');
-        Route::post('/', 'store')->name('store');
-        Route::put('/{propertyUnit}', 'update')->name('update');
-        Route::delete('/{propertyUnit}', 'destroy')->name('destroy');
+        Route::get('/', 'index')->name('index')->middleware('permission:properties,view');
+        Route::post('/', 'store')->name('store')->middleware('permission:properties,create');
+        Route::put('/{propertyUnit}', 'update')->name('update')->middleware('permission:properties,update');
+        Route::delete('/{propertyUnit}', 'destroy')->name('destroy')->middleware('permission:properties,delete');
         Route::get('/dashboard', 'dashboard')->name('dashboard');
     });
